requirements.txt
accelerate==1.12.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asteroid-filterbanks==0.4.0
attrs==25.4.0
audioread==3.1.0
Authlib==1.6.6
av==16.1.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.3
cryptography==46.0.4
ctranslate2==4.6.3
cycler==0.12.1
decorator==5.2.1
deep-translator==1.11.4
distro==1.9.0
dnspython==2.8.0
einops==0.8.2
email-validator==2.3.0
fastapi==0.128.0
faster-whisper==1.2.1
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
groq==1.0.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
julius==0.2.7
kiwisolver==1.4.9
langdetect==1.0.9
lazy_loader==0.4
librosa==0.11.0
lightning==2.6.1
lightning-utilities==0.15.2
llvmlite==0.46.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
onnxruntime==1.23.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
optuna==4.7.0
packaging==26.0
pandas==3.0.0
pillow==12.1.0
platformdirs==4.5.1
pooch==1.9.0
primePy==1.3
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyannote-audio==4.0.3
pyannote-core==6.0.1
pyannote-database==6.1.1
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pyannoteai-sdk==0.4.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMySQL==1.1.2
pyparsing==3.3.2
pyreadline3==3.5.4
pysrt==1.1.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
pytorch-lightning==2.6.1
pytorch-metric-learning==2.9.0
PyYAML==6.0.3
regex==2026.1.15
reportlab==4.4.9
requests==2.32.5
rich==14.3.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
SQLAlchemy==2.0.46
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.8.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.8.0
torchcodec==0.7.0
torchmetrics==1.8.2
tqdm==4.67.2
transformers==5.0.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
yarl==1.22.0
zipp==3.23.0
